doc-powered-rag/requirements.txt
mock==5.1.0
openai==1.41.0
pytest==8.3.2
python-dotenv==1.0.1
Requests==2.32.3
